Extract resolvers and mutations for page
